AdminPerms - Rename files to use kebab case for consistency
